- split query code into util class, add example test code (CDA-47) - apply spotless reformatting to all code
