+++ src/main/java/bio/terra/cda/app/controller/QueryApiController.java
@@ -5,15 +5,14 @@ import bio.terra.cda.generated.controller.QueryApi;
 import bio.terra.cda.generated.model.InlineResponse200;
 import bio.terra.cda.generated.model.QueryNode;
 import bio.terra.cda.service.ping.PingService;
+import java.util.ArrayList;
+import java.util.List;
+import javax.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestParam;
-
-import javax.validation.Valid;
-import java.util.ArrayList;
-import java.util.List;
 
 @Controller
 public class QueryApiController implements QueryApi {
@@ -32,7 +31,8 @@ public class QueryApiController implements QueryApi {
   }
 
   @Override
-  public ResponseEntity<InlineResponse200> booleanQuery(String version, @Valid QueryNode body, @Valid Integer offset, @Valid Integer limit) {
+  public ResponseEntity<InlineResponse200> booleanQuery(
+      String version, @Valid QueryNode body, @Valid Integer offset, @Valid Integer limit) {
     QueryService service = new QueryService();
     // FIXME: need try/catch for error handling.
     final List<String> jsonData = service.runQuery(body);
